refactor: item services are now separate from cource services
