+++ backend/src/modules/courses/controllers/ItemController.ts
@@ -1,23 +1,16 @@
-import {instanceToPlain, plainToInstance} from 'class-transformer';
 import 'reflect-metadata';
 import {
+  JsonController,
   Authorized,
-  BadRequestError,
+  Post,
   Body,
   Get,
-  JsonController,
-  Params,
-  Post,
   Put,
   Delete,
-  HttpError,
+  Params,
   HttpCode,
 } from 'routing-controllers';
-import {CourseRepository} from 'shared/database/providers/mongo/repositories/CourseRepository';
-import {UpdateError} from 'shared/errors/errors';
-import {DeleteError} from 'shared/errors/errors';
-import {Inject, Service} from 'typedi';
-import {Item, ItemsGroup} from '../classes/transformers/Item';
+import {Service, Inject} from 'typedi';
 import {
   CreateItemBody,
   UpdateItemBody,
@@ -28,380 +21,91 @@ import {
   MoveItemParams,
   DeleteItemParams,
 } from '../classes/validators/ItemValidators';
-import {calculateNewOrder} from '../utils/calculateNewOrder';
+import {ItemService} from '../services';
+import {OpenAPI} from 'routing-controllers-openapi';
 
-/**
- * Controller for managing items within course modules and sections.
- * Handles operations such as creation, retrieval, update, and reordering.
- *
- * @category Courses/Controllers
- * @categoryDescription
- * Provides endpoints for working with "items" inside sections of modules
- * within course versions. This includes content like videos, blogs, or quizzes.
- */
-
-@JsonController('/courses')
+@OpenAPI({
+  tags: ['Items'],
+})
+@JsonController('/items')
 @Service()
 export class ItemController {
   constructor(
-    @Inject('CourseRepo') private readonly courseRepo: CourseRepository,
-  ) {
-    if (!this.courseRepo) {
-      throw new Error('CourseRepository is not properly injected');
-    }
-  }
+    @Inject('ItemService') private readonly itemService: ItemService,
+  ) {}
 
-  /**
-   * Create a new item under a specific section of a module in a course version.
-   *
-   * @param params - Route parameters including versionId, moduleId, and sectionId.
-   * @param body - The item data to be created.
-   * @returns The updated itemsGroup and version.
-   *
-   * @throws HTTPError(500) on internal errors.
-   *
-   * @category Courses/Controllers
-   */
-
-  @Authorized(['admin'])
-  @Post('/versions/:versionId/modules/:moduleId/sections/:sectionId/items')
+  @Authorized(['admin', 'instructor'])
+  @Post('/:versionId/:moduleId/:sectionId')
   @HttpCode(201)
+  @OpenAPI({
+    summary: 'Create Item',
+    description: 'Creates a new item within a section.',
+  })
   async create(
     @Params() params: CreateItemParams,
     @Body() body: CreateItemBody,
   ) {
-    try {
-      const {versionId, moduleId, sectionId} = params;
-      //Fetch Version
-      const version = await this.courseRepo.readVersion(versionId);
-
-      //Find Module
-      const module = version.modules.find(m => m.moduleId === moduleId);
-
-      //Find Section
-      const section = module.sections.find(s => s.sectionId === sectionId);
-
-      //Fetch ItemGroup
-      const itemsGroup = plainToInstance(
-        ItemsGroup,
-        await this.courseRepo.readItemsGroup(section.itemsGroupId.toString()),
-      );
-
-      //Create Item
-      const newItem = new Item(body, itemsGroup.items);
-
-      //Add Item to ItemsGroup
-      itemsGroup.items.push(newItem);
-
-      //Update Section Update Date
-      section.updatedAt = new Date();
-
-      //Update Module Update Date
-      module.updatedAt = new Date();
-
-      //Update Version Update Date
-      version.updatedAt = new Date();
-
-      const updatedItemsGroup = await this.courseRepo.updateItemsGroup(
-        section.itemsGroupId.toString(),
-        itemsGroup,
-      );
-
-      //Update Version
-      const updatedVersion = await this.courseRepo.updateVersion(
-        versionId,
-        version,
-      );
-
-      return {
-        itemsGroup: instanceToPlain(updatedItemsGroup),
-        version: instanceToPlain(updatedVersion),
-      };
-    } catch (error) {
-      if (error instanceof Error) {
-        throw new HttpError(500, error.message);
-      }
-    }
+    const {versionId, moduleId, sectionId} = params;
+    return this.itemService.createItem(versionId, moduleId, sectionId, body);
   }
 
-  /**
-   * Retrieve all items from a section of a module in a course version.
-   *
-   * @param params - Route parameters including versionId, moduleId, and sectionId.
-   * @returns The list of items within the section.
-   *
-   * @throws HTTPError(500) on internal errors.
-   *
-   * @category Courses/Controllers
-   */
-
-  @Authorized(['admin', 'instructor', 'student'])
-  @Get('/versions/:versionId/modules/:moduleId/sections/:sectionId/items')
+  @Authorized(['admin', 'instructor'])
+  @Get('/:versionId/:moduleId/:sectionId')
+  @OpenAPI({
+    summary: 'Read All Items',
+    description: 'Retrieves all items from a section.',
+  })
   async readAll(@Params() params: ReadAllItemsParams) {
-    try {
-      const {versionId, moduleId, sectionId} = params;
-      //Fetch Version
-      const version = await this.courseRepo.readVersion(versionId);
-
-      //Find Module
-      const module = version.modules.find(m => m.moduleId === moduleId);
-
-      //Find Section
-      const section = module.sections.find(s => s.sectionId === sectionId);
-
-      //Fetch Items
-      const itemsGroup = await this.courseRepo.readItemsGroup(
-        section.itemsGroupId.toString(),
-      );
-
-      return {
-        itemsGroup: itemsGroup,
-      };
-    } catch (error) {
-      if (error instanceof Error) {
-        throw new HttpError(500, error.message);
-      }
-    }
+    const {versionId, moduleId, sectionId} = params;
+    return this.itemService.readAllItems(versionId, moduleId, sectionId);
   }
 
-  /**
-   * Update an existing item in a section of a module in a course version.
-   *
-   * @param params - Route parameters including versionId, moduleId, sectionId, and itemId.
-   * @param body - Fields to update, including name, description, type, and itemDetails.
-   * @returns The updated itemsGroup and version.
-   *
-   * @throws HTTPError(500) on internal errors.
-   *
-   * @category Courses/Controllers
-   */
-
-  @Authorized(['admin'])
-  @Put(
-    '/versions/:versionId/modules/:moduleId/sections/:sectionId/items/:itemId',
-  )
+  @Authorized(['admin', 'instructor'])
+  @Put('/:versionId/:moduleId/:sectionId/:itemId')
+  @OpenAPI({
+    summary: 'Update Item',
+    description: 'Updates an existing item.',
+  })
   async update(
     @Params() params: UpdateItemParams,
     @Body() body: UpdateItemBody,
   ) {
-    try {
-      const {versionId, moduleId, sectionId, itemId} = params;
-      //Fetch Version
-      const version = await this.courseRepo.readVersion(versionId);
-
-      //Find Module
-      const module = version.modules.find(m => m.moduleId === moduleId);
-
-      //Find Section
-      const section = module.sections.find(s => s.sectionId === sectionId);
-
-      //Fetch ItemsGroup
-      const itemsGroup = await this.courseRepo.readItemsGroup(
-        section.itemsGroupId.toString(),
-      );
-
-      //Find Item
-      const item = itemsGroup.items.find(i => i.itemId === itemId);
-
-      //Update Item
-      Object.assign(item, body.name ? {name: body.name} : {});
-      Object.assign(
-        item,
-        body.description ? {description: body.description} : {},
-      );
-      Object.assign(item, body.type ? {type: body.type} : {});
-
-      //Update Item Details
-      Object.assign(
-        item,
-        body.videoDetails
-          ? {itemDetails: body.videoDetails}
-          : body.blogDetails
-            ? {itemDetails: body.blogDetails}
-            : body.quizDetails
-              ? {itemDetails: body.quizDetails}
-              : {},
-      );
-
-      //Update Section Update Date
-      section.updatedAt = new Date();
-
-      //Update Module Update Date
-      module.updatedAt = new Date();
-
-      //Update Version Update Date
-      version.updatedAt = new Date();
-
-      //Update ItemsGroup
-      const updatedItemsGroup = await this.courseRepo.updateItemsGroup(
-        section.itemsGroupId.toString(),
-        itemsGroup,
-      );
-
-      //Update Version
-      const updatedVersion = await this.courseRepo.updateVersion(
-        versionId,
-        version,
-      );
-
-      return {
-        itemsGroup: instanceToPlain(updatedItemsGroup),
-        version: instanceToPlain(updatedVersion),
-      };
-    } catch (error) {
-      if (error instanceof Error) {
-        throw new HttpError(500, error.message);
-      }
-    }
+    const {versionId, moduleId, sectionId, itemId} = params;
+    return this.itemService.updateItem(
+      versionId,
+      moduleId,
+      sectionId,
+      itemId,
+      body,
+    );
   }
 
-  /**
-   * Delete an item from a section of a module in a course version.
-   * @param params - Route parameters including versionId, moduleId, sectionId, and itemId.
-   * @return The updated itemsGroup and version.
-   * @throw HTTPError(500) on internal errors.
-   * @category Courses/Controllers
-   */
-
-  @Authorized(['instructor', 'admin'])
-  @Delete('/itemGroups/:itemsGroupId/items/:itemId')
+  @Authorized(['admin', 'instructor'])
+  @Delete('/:itemsGroupId/:itemId')
+  @OpenAPI({
+    summary: 'Delete Item',
+    description: 'Deletes an item from the items group.',
+  })
   async delete(@Params() params: DeleteItemParams) {
-    try {
-      const {itemsGroupId, itemId} = params;
-
-      if (!itemsGroupId || !itemId) {
-        throw new DeleteError('Missing required parameters');
-      }
-
-      //Fetch ItemsGroup
-      const itemsGroup = await this.courseRepo.readItemsGroup(itemsGroupId);
-      if (!itemsGroup) {
-        throw new DeleteError('ItemsGroup not found');
-      }
-
-      const itemToDelete = itemsGroup.items.find(
-        i => i.itemId.toString() === itemId,
-      );
-
-      if (!itemToDelete) {
-        throw new DeleteError('Item not found');
-      }
-
-      const deletionStatus = await this.courseRepo.deleteItem(
-        itemsGroupId,
-        itemId,
-      );
-
-      if (!deletionStatus) {
-        throw new Error('Unable to delete item');
-      }
-
-      const updatedItemsGroup =
-        await this.courseRepo.readItemsGroup(itemsGroupId);
-
-      return {
-        deletedItem: instanceToPlain(itemToDelete),
-        updatedItemsGroup: instanceToPlain(updatedItemsGroup),
-      };
-    } catch (error) {
-      if (error.message === 'Item not found') {
-        throw new HttpError(404, error.message);
-      }
-      if (error.message === 'Missing required parameters') {
-        throw new BadRequestError(error.message);
-      }
-      if (error instanceof Error) {
-        throw new HttpError(500, error.message);
-      }
-    }
+    const {itemsGroupId, itemId} = params;
+    return this.itemService.deleteItem(itemsGroupId, itemId);
   }
 
-  /**
-   * Move an item to a new position within a section by recalculating its order.
-   *
-   * @param params - Route parameters including versionId, moduleId, sectionId, and itemId.
-   * @param body - Movement instructions including `afterItemId` or `beforeItemId`.
-   * @returns The updated itemsGroup and version.
-   *
-   * @throws BadRequestError if both afterItemId and beforeItemId are missing.
-   * @throws HTTPError(500) on internal errors.
-   *
-   * @category Courses/Controllers
-   */
-  @Authorized(['admin'])
-  @Put(
-    '/versions/:versionId/modules/:moduleId/sections/:sectionId/items/:itemId/move',
-  )
+  @Authorized(['admin', 'instructor'])
+  @Put('/move/:versionId/:moduleId/:sectionId/:itemId')
+  @OpenAPI({
+    summary: 'Move Item',
+    description:
+      'Reorders an item within its section by moving it before or after another item.',
+  })
   async move(@Params() params: MoveItemParams, @Body() body: MoveItemBody) {
-    try {
-      const {versionId, moduleId, sectionId, itemId} = params;
-      const {afterItemId, beforeItemId} = body;
-
-      if (!afterItemId && !beforeItemId) {
-        throw new UpdateError('Either afterItemId or beforeItemId is required');
-      }
-
-      //Fetch Version
-      const version = await this.courseRepo.readVersion(versionId);
-
-      //Find Module
-      const module = version.modules.find(m => m.moduleId === moduleId);
-
-      //Find Section
-      const section = module.sections.find(s => s.sectionId === sectionId);
-
-      //Fetch ItemsGroup
-      const itemsGroup = await this.courseRepo.readItemsGroup(
-        section.itemsGroupId.toString(),
-      );
-
-      //Find Item
-      const item = itemsGroup.items.find(i => i.itemId === itemId);
-
-      //Sort Items based on order
-      const sortedItems = itemsGroup.items.sort((a, b) =>
-        a.order.localeCompare(b.order),
-      );
-
-      //Calculate New Order
-      const newOrder = calculateNewOrder(
-        sortedItems,
-        'itemId',
-        afterItemId,
-        beforeItemId,
-      );
-
-      //Update Item Order
-      item.order = newOrder;
-
-      //Change the updatedAt dates
-      section.updatedAt = new Date();
-      module.updatedAt = new Date();
-      version.updatedAt = new Date();
-
-      //Update ItemsGroup
-      const updatedItemsGroup = await this.courseRepo.updateItemsGroup(
-        section.itemsGroupId.toString(),
-        itemsGroup,
-      );
-
-      //Update Version
-      const updatedVersion = await this.courseRepo.updateVersion(
-        versionId,
-        version,
-      );
-
-      return {
-        itemsGroup: instanceToPlain(updatedItemsGroup),
-        version: instanceToPlain(updatedVersion),
-      };
-    } catch (error) {
-      if (error instanceof UpdateError) {
-        throw new BadRequestError(error.message);
-      }
-      if (error instanceof Error) {
-        throw new HttpError(500, error.message);
-      }
-    }
+    const {versionId, moduleId, sectionId, itemId} = params;
+    return this.itemService.moveItem(
+      versionId,
+      moduleId,
+      sectionId,
+      itemId,
+      body,
+    );
   }
 }
