+++ backend/src/modules/courses/services/index.ts
@@ -1,1 +1,2 @@
 export * from './CourseService';
+export * from './ItemService';
