+++ backend/src/shared/database/index.ts
@@ -1,3 +1,4 @@
 export * from './interfaces/IDatabase';
 export * from './interfaces/IUserRepository';
 export * from './interfaces/ICourseRepository';
+export * from './interfaces/IItemRepository';
